[ML] Handle cache size too small for single item
The LFU cache was assuming that if a new item wouldn't
fit into it then there must already be a cached item
in it. This wasn't true in the case of a very tiny
cache size.

The code now accounts for the possibility of an item
not fitting into an empty cache.

This scenario was always very unlikely in production,
but more likely for test scenarios.

Fixes #2370