[ML] Improve quantile estimation accuracy and median accuracy for anomaly detection (#2382)
We use a histogram sketch for estimating data quantiles and also computing the time bucket median for anomaly
detection.

Once data are merged into buckets we have to make some assumptions about how values are distributed on buckets.
Previously, we assumed data are uniformly distributed on buckets whose endpoints are the midpoints between the
bucket "centres" we track. In fact, the points we track are the weighted means of the values in each bucket, so it is
more accurate to assume that roughly half the data in a bucket falls either side of this point.

This changes the interpolation scheme to use the same endpoints but to incorporate this new assumption. This gives
a very nice additional property: a quantile which falls between two buckets each containing a single value is computed
exactly. An immediate corollary is that all quantiles are exact if the data size is less than the sketch size. Previously, we
used piecewise constant interpolation for estimating the median because it is exact in this case. We now cut over to
linear interpolation. This is attractive because for large data, if the data distribution is smooth, linear interpolation is
significantly more accurate for fixed memory usage than piecewise constant interpolation.