[ML] Boosted tree row split calculation performance (#2380)
This focuses on refreshSplitsCache whose cost is dominated by std::upper_bound. clang in particular generates
poor assembly for std::upper_bound for our usage: creating an (unpredictable) branch instead of a conditional
move on the hot path. In any case, I wanted to to see if we can do better here using a search tree strategy.

The search tree can potentially benefit from better LOR when using branching factor greater than 2 because we
can pack split values into consecutive memory. Another advantage is we always compile with SSE4.2 so we have
single instructions for doing the comparison of 4 distinct values. The resulting data structure is implemented as
CSearchTree. (Note that this is in CBoostedTreeUtils which was getting included more than it should because it
contained functionality which wasn't really logically part of it anymore. I moved this into the correct places and
cutdown the includes as a result.)

On synthetic benchmarks, with uniform random data, pretty much the worst case for the previous implementation,
I get around a 4.5X speedup vs std::upper_bound for our typical usage set sizes. For the benchmark which motivated
this change, I cut the cost of refreshSplitsCache by a little over 2X.