[ML] Detect debug-only compile errors in PR builds (#2389)
People rarely build the code with debug enabled, so it's
possible for an error in the code that only affects debug
builds to get committed.

This change adds a step to build (but not unit test) the
code with debug enabled in every PR build. Doing this in
the cross compile sub-job should mean that it doesn't
increase the overall PR CI duration, as the cross compile
sub-job is faster than the sub-jobs that run unit tests.