[ML] Improve memory management for training on large data sets (#2388)
Profiling showed we spend a surprising amount of time creating and destroying leaf statistics state training on
large data sets. These objects are large O("number splits" * "number features" * ("gradient" + "hessian")) we
also have to map out the memory of individual split vectors and this work has to be done each time we allocate
a new object. In fact, we can just recycle the objects when we determine that no splits are useful at a leaf as
long as the objects are "conformable", i.e. same dimensions, which they usually are.

This change saves up to 40% of the work on the main thread for some data sets, it should also lead to better
LOR.