[ML] Improve the layout of aggregate derivatives for training (#2391)
This migrates the count of values we maintain for splits into the buffer we use for storing accumulated derivatives.
This has the advantage that we can reset and copy it in the same operations we use for derivates rather than
separately iterating over the collection to update counts. Furthermore, we often get these operations for free
since we pad buffers to maintain alignment.

This change showed up some deficiencies in CChecksum particularly for pointers. I fixed these and pulled non
template implementations out of the header since it is widely included.