[ML] Raise the limit on maximum number of classes for multi-class classification (#2395)
This addresses the bottleneck preventing us from increasing the maximum number of distinct classes we'll
support and raises the limit from 30 to 100.

The issue is computing the hessian which scales like n(n+1)/2 for n distinct classes. Here I've taken the
approach of targeting just the classes which satisfy arg top-k_{c \in C}{ \sum_i 1{c_i = c} loss(p_i, c_i) }
where p_i are the prediction probabilities for the i'th example, c_i is the true class, C is the set of all
classes and 1{.} denotes the indicator function, when computing best splits. This limits the hessian to
k(k+1)/2 and I've capped k at 20. We simply lump all classes not selected into a hold all other class when
we chose the best splits and compute the expected derivatives of the loss for them. However, since we
choose classes afresh for each round of boosting, we should still find splits which discriminate between
all classes in the forest as a whole.

Closes #2246.