[8.4] [ML] Don't keep categorization tokens when existing category matches
In categorization, when a new message matches an existing category
we don't need to keep a record of any new distinct tokens found in
the new message, as the category definition doesn't need them.

We used to store all unique tokens ever seen, but this can lead to
unbounded memory build-up if a particular type of logs has some
sort of unique ID that doesn't get filtered by the token filter of
the categorization analyzer.

This PR removes new distinct tokens from the token store for all
code paths other than a message causing a new category to be
created.

Backport of #2398