[ML] Improvements to the outlier influence calculation (#2256)
Currently, we compute influence of features on the outlier score by searching for the minimum
value of the outlier score allowing one coordinate to vary at a time. This works well if the data
aren't projected, but when the data are high dimensional we randomly project them. We then
have to work out how changes in projected coordinates should be attributed to the original
features. This is only possible approximately and can lead to confusing results.

The normal definition of influence is the (Gateaux) derivative of the outlier score with respect
to the feature values. This has the advantage that we can simply compute the change in the
projection space for small changes in a single feature value. This changes switches to use this
definition. For the unprojected case the results are broadly similar: the unit tests pass with minor
tweaks. However, this fixes confusing results in the other case.