[ML] Fix for 'No statistics' error message (#2410)
An anomaly detector job ignores records that fall in the initial, incomplete bucket. However, a model is created with a start time of -1 in the expectation that more records will be received that fall into subsequent buckets. In the case that no more records are received the model continues to exist with start time of -1.

In the scenario that a job is created initially as lookback and has such models with start time of -1 in existence at the point the job closes, and then is re-opened as a realtime job at some point in the future, any attempt to sample the model will fail with the 'No statistics at ...' error message due to the start time (-1) falling outside the current bucket.

This change always updates the model's current bucket stats start time to that of the current record, even if the record data is not to be added to the associated data gatherer.