[ML] Fix for "No counts available" error message
When restarting a look-back job as a real-time one on occasion error
messages similar to
'[CBucketGatherer.cc@464] No counts available at 1585698300, current bucket = [1585699200,1585700100)'
are seen.

Investigation reveals that these error messages stem from the situation
where an incomplete initial bucket for a partition was persisted at the
close of the look-back job. When the job is re-opened as real-time the
start time of the anomaly detector in question is still before that of
the associated bucket gatherer, triggering the error.

This PR adds a check to determine if an incomplete initial bucket has
been restored from persisted state, and if so to skip any attempt to
output results for that bucket - which is essentially what would have
happened prior to this change except without logging the error.

Fixes #2411