[ML] Fix instructions for building PyTorch on Windows
On Windows, recent updates of Visual Studio 2019 include an
experimental improved OMP library. Because this library is
experimental redistribution is not permitted.

When we built PyTorch 1.13.1 CMake's FindOpenMP module decided
to link to this experimental OMP library, which then means the
resulting library only works on machines with Visual Studio
installed.

It's not very clear where in the PyTorch build configuration
the fault lies. It could be in CMake itself or maybe there's
some option PyTorch could be using to control CMake more
tightly. CMake was _also_ linking to Microsoft's GA OMP
library, so this PR works around the problem by temporarily
renaming the experimental libraries during the build so that
CMake does not find them.

Fixes #2432