[ML] Improve handling of time shifts such as daylight savings (#2479)
We received a data set where we had obvious deficiencies adjusting to daylight saving. There were two
problems:
1. If the time series flat lines, say over a weekend, we can't necessarily segment into "before" and "after"
    the time shift and this was preventing us detecting the shift
2. We reverted shifts we applied sometimes where the time series is flat because in this case the test
    we use has low power but can have high significance.
For case 1 we simply need to check if a non-zero time shift over the full window has high significance.
For case 2 we can use the same idea to revert we use when identifying whether to shift: check that we
see values capturing a reasonable proportion of the time series variation before we trust the result.