[ML] Migrate to Backstage for BuildKite pipeline definitions
Add a "catalog-info.yaml" file that contains definitions of the ml-cpp
BuildKite pipelines. These will take precedence of those defined in the
CI repo. Once the catalog is merged to main and detected by the CI
Backstage system the old ml-cpp* pipeline definition files in the CI
repo can be deleted.