[ML] BuildKite fix linux x86_64 dependencies (#2508)
The dependency libraries for the linux x86_64 build are not being collected correctly. An error in the build log indicates that this is due to incorrect interpretation of the CPP_CROSS_COMPILE environment variable.

This PR now checks if the CPP_CROSS_COMPILE environment variable is defined or the empty string.