[ML] Improve calendar cyclic component modelling for long bucket lengths (#2493)
Our test to find important calendar components to model was not reliable for long bucket lengths.
This change addresses some of the short comings:
1. We take too long to warm up the error quantiles,
2. We can't calculate a p-value per repeat because we don't get to observe enough errors per
    repeat to estimate this properly,
3. We can't estimate high error percentiles accurately with long buckets. (Instead we switch to
    counting events with fixed expected count rather than fixed probability for long bucket length.)
We also now estimate a lower bound on the right-tail probability for the feature directly from its
sum error. This can fall between the error distribution points we count and improves the test
sensitivity. Finally, it is useful to be able to detect multiple components per test, otherwise it can
take a long time to initialise every component which is needed. This adds the ability to initialise
multiple statistically significant components per test run.