Adjust CEventRateModelTest tests to account for changes to Boost unordered containers
Changes to Boost unordered_map have resulted in the order of retrieval of elements being different across versions, i.e. identical collections of elements will be stored in a different order in a Boost 1.77 unordered_map as opposed to a Boost 1.83 map. Both orders are legitimate and deterministic.

For the test cases in question probability and influences are being calculated by CProbabilityAndInfluenceCalculator::calculate which retrieves influencers & probability pairs from an unordered map, stores them in a vector and then sorts that vector based on the probability values. However, as the probabilities of the influencers are identical the sort has no effect.