[ML] Build Linux x86_64 image in VM
Up to now we've built the Linux x86_64 artifacts in a Docker
container hosted on Kubernetes. This means the BuildKite
agent needs to run in the build Docker image and use its
Git to get the correct code to build. This no longer works
with the ancient version of Git on CentOS 6. By building
in a VM, BuildKite can run on the (modern) VM distribution
and just the compiler needs to run in the ancient Docker
image (which is needed for 7.x platform support).