[ML] Better fallback when hardware_concurrency fails for pytorch inference (#2607)
The advice is to treat this as a hint only. It will return 0 if it can't be computed for the platform
and in this case we should size the thread pool for the requested number of allocations.

Also, we were never actually outputting how many allocations a pytorch_inference process has
available so if std::thread::hardware_concurrency() fails then we don't get to find out.