[ML] Upgrade MKL to version 2024.0
After upgrading PyTorch to version 2.1.2 we encountered
crashes when using the optimised version of ELSER v2 on
Linux x86_64. The problem appeared to be caused either by
a bug in the MKL version 2020.4 or else an incompatibility
between that version of MKL and PyTorch 2.1.2. An upgrade
to MKL version 2024.0 appears to resolve it.

Fixes #2617