[ML] Suppress LNK4217 linker warning for libMlCore
The link step for libMlCore generates many LNK4217 warnings referencing
Boost::Json symbols. The warning is indicating that the definitions of
the symbols have already been imprted in othe translation units in the
library.

These warnings are harmless and are safe to suppress (and are not
avoidable by any other means as far as I can tell)