[ML] Use Orka for macOS ARM builds. (#2662)
At the moment macOS ARM builds are performed on bare metal Mac Minis
provided by MacStadium. For consistency with the rest of engineering,
ease of maintenance, reproducibility of builds, automation of provisioning
builders etc. it is preferred that we migrate to using VMs provided by
MacStadium Orka.

This PR makes use of an image built using `packer`. See #2663 for the associated `packer` config and scripts.