[ML] Fix error introduced by backport (#2671)
The backport of the Orka related changes to the 7.17 branch introduced a
change that inadvertently set the build type to be `DEBUG`. (This is due
to differences in the build scripts in 7.17 to those more recent
branches). Consequently the build takes ~ 3.5 hours to complete and all
binaries will have full debug symbols and assertions enabled.

Merging without review due to the urgency of this fix.