[ML] Remove Python code folder #2595
This PR removes the folder jupyter and all references to it. The folder contained Python code for rapid prototyping of new functionality, mainly incremental learning. As we didn't use this functionality for a while, I removed it to reduce the maintenance overhead.

As this is not a customer-facing functionality, I mark it as a non-issue.