[ML] Add `codesign` CMake target (#2683)
This PR adds the `codesign_autodetect` and `codesign_pytorch_inference`
targets to CMake. These targets add the `get-task-allow` entitlement to
the respective executables, which is necessary to allow debugging with
e.g. `Instruments` on `macOS`

The `codesign` targets are only applicable on `macOS`. Attempting to run
them on any other platform will fail with an explanatory message.