[ML] Create a Docker build image using the latest PyTorch stable branch. (#2685)
On Linux x86_64, provide the ability to create Docker build images that use the code pulled from the HEAD of the PyTorch `viable/strict` branch.

Upon successful build and push of such a Docker image, trigger a build of the ml-cpp code in it.

Once the ml-cpp build has succeeded it in turn should trigger a pipeline in the QAF repo that runs a set of tests exercising the pytorch_inference executable.

To make this possible a new pipeline - ml-cpp-pytorch-build - is required. This is defined in the catalog-info.yaml file and won't be created until catalog-info.yaml is merged to main and backstage magic does its thing.

Some tweaks to our existing Buildkite framework have been made in order that existing code can be better be re-used, so that e.g. just a linux x86_64 build step can be dynamically created.

TBD: 
 * The name of the QA PyTorch testing pipeline is required.
 * The creation of a machine user to push docker images from within a buildkite job.