[ML] Persist callback detection rules in the detector model state (#2739)
Some of the detection rules, e.g. callbacks, should be executed only once. Usually, we would not persist and restore detection rules, relying on the fact that they are listed in the job configuration.

However, to prevent a rule like forcing a time shift from being executed a second time if the job is restored from a snapshot, we need to persist with the list of rules that were applied already.

Since CAnomalyDetectorModel has no state that is persisted/restored, I moved the vector with the checksum into the classes CCountingModel, CPopulationModel, and CIndividualModel.

The logic for checking/marking applied rules is implemented in CDetectionRule::executeCallback.