[ML] Correct handling of config updates (#2821)
* Enable updating of custom rules
* Correctly update filters (scoped rules)

These changes fix a bug whereby memory corruption was caused when updating filters for an open job. The bug had previously been masked by code that reset the job's analysis config to its original state after every config update - the downside of this was that custom rules could not be updated on the fly, only by stopping and restarting the job.