[ML] Fix 3rd party lib depends for Win 2022 builds
The CMake configuration was failing to locate some vital 3rd party
libraries and include them for redistribution for Windows 2022 builds.
This isn't an issue for our internal machines for building and testing
as those files are already present there, but is very much an issue for
non-build machines that are missing the required libraries.