[8.19] [ML] Restrict file system access for pytorch models (#2851) (#2853)
* [ML] Restrict file system access for pytorch models (#2851)

This PR ensures that the PyTorch models are not allowed to access the file system. It accomplishes the goal by inspecting the model's operations and prohibiting the loading of models with operations that read or write files.

(cherry picked from commit fff4062ad42ca40d9cb765505c83f580b72b106b)

* rewrite for C++14 compatibility