[ML] Add a script to run each unit test separately
Add a script to provide a wrapper around the call to "cmake" that runs the test cases and provides some flexibility as to how the tests should be run in terms of how they are spread across processes. This is necessary when trying to isolate the impact memory usage of tests have upon one another.