[ML] Report the "actual" memory usage of the autodetect process (#2846)
Determine the actual memory usgae of the autodetect process as reported by the OS, e.g. on Linux this mould be the value of the maximum resident set size returned by a call to `getrusage`.

Add this value to the model size stats record returned to the ES Java process so it can be included in the `job counts` tab for anomaly detection jobs.

---------

Co-authored-by: Valeriy Khakhutskyy <1292899+valeriy42@users.noreply.github.com>