[ML] Bump PyTorch version to 2.7.1 (#2882)
PyTorch 2.7.1 contains essential security fixes. We need to upgrade our build images accordingly.

This initial commit contains the changes for the Linux build images. macOS and Windows images willbe updated in due course.

This is essentially a backport of #2863 to the 8.19 branch.

Note, Windows is still using the older PyTorch version.